double_check is gone for good

+++ app/controllers/act_memberships_controller.rb
@@ -10,8 +10,7 @@ class ActMembershipsController < ApplicationController
 		user = User.find_by_id(params[:joining_user])
 		act_id = params[:offering_id]
 
-			return unless double_check {
-		name_is_valid?(user, act_type) }
+		if !name_is_valid?(user, act_type); raise Errors::FlowError.new; end
 
 		acts_membership = user.send("#{act_type}s_membership")
 		joining_act = act_type.camelize.constantize.find_by_id(act_id)
@@ -33,8 +32,7 @@ class ActMembershipsController < ApplicationController
 		user = User.find_by_id(params[:leaving_user])
 		act_id = params[:offering_id]
 
-			return unless double_check {
-		name_is_valid?(user, act_type) }
+		if !name_is_valid?(user, act_type); raise Errors::FlowError.new; end
 
 		acts_memberships = user.act_memberships.where(act_type: act_type, act_id: act_id)
 		acts_memberships.each.destroy unless acts_memberships == []
